// client/src/models/events.d.ts
export namespace Events {
  export interface IncomingParams {
    id?: ID;
    stockSelectorCallback?: (stockId: ID) => void;
  }
  
  export interface ActionParams {
    selectorCallbackSetter?: (selectorCallback: { selectorCallback: (stockId: ID) => void }) => void;
    idStateSetter?: (id: ID) => any;
    displayStateSetter?: (display: boolean) => any;
    refreshState?: number;
    refreshStateSetter?: (refreshIteration: number) => any;
    searchInputSetter?: (inputVal: string) => void;
  }
// ...
}